Add files via upload
Using Sentiment Analysis and Data Visualization to explore BBNaija2020.
I recently worked on a side project with the premises that if people on Twitter
actually vote based on how they talk about the housemates in their tweets.
What happens if people vote as they tweet? Of course, they don’t but what if
they do? What if there is an AI that reads your tweet on twitter and could
categorize it into Positive, negative and neutral and use this as a basis for who
gets evicted and who doesn’t?
These are the questions that myself and @prommy15 tried to solve on this
project.
Disclaimer: This project is just a project. It is not sponsored by anyone and
neither is it directed towards anymore. This is simply a sentiment analysis and
visualization of tweets, not a deep analysis and might not in any way affect
the ACTUAL votes.
Using the housemates’ names as hashtags, 1000 tweets were collected for
each housemates from Sunday to Thursday. We broke it into 3 categories
based on how people directed the tweets; Positive, Negative and Neutral.
Positive tweets: Tweets that says something positive about an housemate.
Negative tweets: Tweets that says something negative about a housemate.
Neutral: Tweets are neither positive nor negative.
Based on the premise that
 people with neutral comments about an housemate are indecisive of where
they stand concerning a housemate.
 People with positive comments are more likely to vote for such
housemate.
 People with negative comments are less likely to vote for such housemate.
We further simplified the data as shown below
By neglecting tweets that are neutral because they are less likely to vote for
any housemate, Positive comments as a counter for those that are likely to
vote for a housemate and Negative for those that are likely not to vote for a
housemate.
By using the relationship
Possible eviction = positive-negative to get the value for least and most likely
to be evicted in the BBnaija.
According to the tweets from Sunday to Thursday, the results is as shown
below with Praise leading the chart and Dora among the last 4.
Please note that this visualization is only valid if tweets are considered votes
but graciously they are not, you have to actually vote online.
Based on this, Neo is most likely to be among the bottom 4 if nothing changes
between now and Saturday.
This is simply a visualization of data scrapped off Twitter for Educational
purposes, it is no way directed at anyone nor is it an indication of how well
you have voted. This is just a personal project based on what is going on right
now and IT IS NOT SPONSORED BY ANYBODY NOR ANY BODY.
Who else can’t wait for Ebuka to reveal the votes tomorrow night and who is
looking forward to see what next week tweets will look like?
Data Scrapped, analyzed and visualized: by me(https://mobile.twitter.com/bamidezy/status/1300056431660593154?s=19)
Information designer: https://mobile.twitter.com/Prommy15